delete confirmation etc

+++ src/app/admin-manage-work/admin-manage-work.component.ts
@@ -1,5 +1,7 @@
 import {Component, OnInit} from '@angular/core';
 import {DataService} from "../data.service";
+import {DeleteConfirmationComponent} from "../delete-confirmation/delete-confirmation.component";
+import {MatDialog} from "@angular/material/dialog";
 
 @Component({
   selector: 'app-admin-manage-work',
@@ -10,7 +12,7 @@ export class AdminManageWorkComponent implements OnInit {
 
   work = [];
 
-  constructor(private dataService: DataService) {
+  constructor(private dataService: DataService, private dialog: MatDialog) {
   }
 
 
@@ -23,11 +25,23 @@ export class AdminManageWorkComponent implements OnInit {
   }
 
   deleteWork(id: number) {
-    this.dataService.deleteWork(id).subscribe({
-      next: () => {
-        setTimeout(() => {
-          window.location.reload();
-        }, 1000);
+    const dialogRef = this.dialog.open(DeleteConfirmationComponent, {
+      data: {
+        title: 'Conferma eliminazione',
+        message: 'Sei sicuro di voler eliminare questo intervento?'
+      }
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.dataService.deleteWork(id).subscribe({
+          next: (data: any) => {
+            console.log(data);
+            setTimeout(() => {
+              window.location.reload();
+            }, 1000);
+          }
+        });
       }
     });
   }
